requirements.txt
asgiref==3.7.2
astroid==3.0.0
dill==0.3.7
Django==4.2.6
django-cors-headers==4.2.0
django-filter==23.3
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
install==1.3.5
isort==5.12.0
Markdown==3.4.4
mccabe==0.7.0
packaging==23.2
Pillow==10.0.1
platformdirs==3.11.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pylint==3.0.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
uritemplate==4.1.1
